content: move blog.py + legal.py into trusted_router/content/ package (#42 step 1)
Audit finding: ~1.6k lines of marketing/legal content lived at the top level of
the billing control plane. Group them in a dedicated content/ package so a
content edit has a small, clearly-scoped blast radius (and a future move to a
separate deployable is a directory move, not a package-wide hunt). Pure module
relocation via git mv — blog.py/legal.py use absolute imports, unaffected.
Updated the 3 importers (dashboard.py x2, scripts/generate_blog_og.py).
dashboard.py itself (marketing rendering) is a larger follow-up.
