+++ src/trusted_router/content/__init__.py
@@ -1,0 +1,6 @@
+"""Marketing / legal content, isolated from the billing control-plane modules.
+
+Blog posts, legal packets, and subprocessor data live here so a content edit has
+a small, clearly-scoped blast radius (and so a future move to a separate
+deployable is a directory move, not a hunt through the top-level package).
+"""
